got the pipeline working on mps, made a number small improvements on output
added model_print_summary
renamed the model train() eval() and test() functions to set_train_mode() ... because pytoerch already has built-in train() functions that have a specific method signature
added input shape calculation and printing model summary architecture

had to modify the pipeline so that the precision is set based on config, use float32 for mps devices!

+++ preprocessing/dataset.py
@@ -71,6 +71,7 @@ class LOBSequenceDataset(Dataset):
         # Move to device if specified
         if self.device:
             x_tensor = x_tensor.to(self.device)
-            y_tensor = y_tensor.to(self.device)
+            y_tensor = y_tensor.to(self.device, dtype=torch.long)
+            # at least on mps, f1 score requires conversion to long / int64
 
         return x_tensor, y_tensor
